Extract `SurveyQuestion` component.
